+++ burner/extension/package.json
@@ -1,14 +1,15 @@
 {
   "name": "@starknet/burner",
   "private": false,
-  "version": "0.3.5",
+  "version": "0.3.6",
   "description": "Starknet Burner",
   "type": "module",
   "module": "./src/index.js",
   "types": "./dist/index.d.ts",
+  "unpkg": "./dist/index.umd.js",
   "exports": {
     ".": {
-      "require": "./dist/index.umd.cjs",
+      "require": "./dist/index.umd.js",
       "import": "./dist/index.js"
     }
   },
