+++ burner/extension/src/lib/3.x/inpage/interface.ts
@@ -15,6 +15,7 @@ export interface IStarknetWindowObject {
   isConnected: boolean;
   selectedAddress?: string;
   chainId?: StarknetChainId;
+  compatible: string;
 
   provider?: ProviderInterface;
   account?: AccountInterface;
