+++ burner/extension/src/lib/3.x/inpage/signer.ts
